+++ learn/profile-cache.ts
@@ -4,7 +4,7 @@ import core from '../chat/core';
 import {Character as ComplexCharacter, CharacterGroup, Guestbook} from '../site/character_page/interfaces';
 import { AsyncCache } from './async-cache';
 import { Matcher, MatchReport, Scoring } from './matcher';
-import { PermanentIndexedStore } from './store/sql-store';
+import { PermanentIndexedStore } from './store/types';
 import { CharacterImage, SimpleCharacter } from '../interfaces';